Make helper methods in classifiers static
